process metrics

+++ models/cluster.go
@@ -69,8 +69,8 @@ type Hz struct {
 }
 
 type PageCache struct {
-	LogHitRate     int `json:"log_hit_rate"`
-	StorageHitRate int `json:"storage_hit_rate"`
+	LogHitRate     float64 `json:"log_hit_rate"`
+	StorageHitRate float64 `json:"storage_hit_rate"`
 }
 
 type RecoveryState struct {
